% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{add_range}
\alias{add_range}
\title{add_range - add a range to a value}
\usage{
add_range(x, bound)
}
\arguments{
\item{x}{data to enter. must be numeric}

\item{bound}{range of data to enter}
}
\description{
This builds a range around a value
}
