% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_add.R
\name{consecutive_add}
\alias{consecutive_add}
\title{consecutive_add - add consecutive events}
\usage{
consecutive_add(data, threshold, use_index = F, index_value = NULL)
}
\arguments{
\item{data}{the a data frame or vector to count}

\item{threshold}{what value should the distinguishing be?}

\item{use_index}{boolean, should it index a value for start and end?}

\item{index_value}{if use_index true, provide a value for labels}
}
\description{
This function will count repeat observations and add them together
it creates a new data frame for duration of each event
}
