% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{exit}
\alias{exit}
\title{exit - a check to exit looping functions}
\usage{
exit(input)
}
\arguments{
\item{input}{a text character - should be from a while loop with user input}
}
\description{
exit - a check to exit looping functions
}
