% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{in_range}
\alias{in_range}
\title{in_range - check to see if value is in a range}
\usage{
in_range(x, bounds)
}
\arguments{
\item{x}{data to check}

\item{bounds}{either a single value or two points}
}
\description{
checks if values are in a range
}
