% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_r.R
\name{mult_r}
\alias{mult_r}
\title{mult_r - a function to source multiple files at once}
\usage{
mult_r(path, files, entry)
}
\arguments{
\item{path}{the folder which contains all your files}

\item{files}{which files to enter from the folder}

\item{entry}{the enter function - either load or source}
}
\description{
mult_r - a function to source multiple files at once
}
