% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_csv.R
\name{read_multiple_csv}
\alias{read_multiple_csv}
\title{read_multiple_csv - read multiple csv files}
\usage{
read_multiple_csv(path, index, header = TRUE, as.is = TRUE)
}
\arguments{
\item{path}{the folder you want to pull from}

\item{index}{the index of all csv files in the working directory}

\item{header}{to talk to read.csv, default true}

\item{as.is}{to talk to read.csv, default true}
}
\description{
This function will take multiple csv files
It will read them and return all as one dataframe
It is it will read files in the order of provided index
}
